Add solution for asynchronous JSON file processing test
In this commit, we have implemented a solution for an asynchronous JSON file processing test. The solution involves loading the JSON file asynchronously and calculating the average of numbers. This solution is documented in a separate markdown (MD) file while the synchronous version in the JsonFileTest.cs file has been replaced with a TODO comment for an async version.